requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-admin-sortable2==2.1.10
django-cors-headers==4.9.0
django-vite==3.1.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
pylint-django==2.6.1
pylint==3.3.9
